swap rotate done

+++ libs/libft/srcs/ft_strdup.c
@@ -5,7 +5,7 @@ char	*ft_strdup(const char *s1)
 	int		len;
 	char	*scopy;
 
-	len = ft_strlen(s1);
+	len = (int)ft_strlen(s1);
 	scopy = malloc(len + 1);
 	if (scopy == NULL)
 		return (NULL);
